Fixed a Bug Within the Queue Index Logic
If a song version was different from what was in the playlistQueue index, it would cause the queue to not move. Spotify does a thing called version change where the song that's playing is different from the song you played under the hood. However, It is still the same song its just the song data that's completly different.