Fixed a Playback bug
If the user had a track on repeat and clicked play skip on a new track, it would play the new track but not index the queue and keep the new song on repeat. This was not intended. It now plays the new song turns repeat song off and indexes the queue.